pyproject.toml
[tool.poetry]
name = "tdg-mvp"
version = "0.1.0"
description = "Test Data Generator from OpenAPI specs"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
python = "^3.11"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
jinja2 = "^3.1.3"
httpx = "^0.25.2"
pyyaml = "^6.0.1"
jsonschema = "^4.20.0"
faker = "^22.0.0"
aiofiles = "^23.2.1"
openai = "^1.7.0"
anthropic = "^0.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
ruff = "^0.1.11"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]